+++ index.js
@@ -1,0 +1,16 @@
+const GoogleSheet = require('./googleSheets');
+GoogleSheet.getRow();
+
+let rows = [{
+    email: 'test1@gmail.com',
+    user_name: 'test1',
+    phone: '0512344121',
+    password: '!@$$!@E'
+},{
+    email: 'test2@gmail.com',
+    user_name: 'test2',
+    phone: '0513453242',
+    password: '!&^%$^##'
+}];
+
+//GoogleSheet.addRow(rows);
